backend complete

+++ server/routes/user.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const dbUtils = require("../services/databaseUtils");
+const dbUtils = require("../utility/databaseUtils");
 const User = require("../models/user.js");
 
 
@@ -9,6 +9,8 @@ const User = require("../models/user.js");
 /* GET user by id. */
 router.get("/:userId", async function (req, res) {
 	const userId = req.params.userId;
+  const currUser = res.locals.userData;
+  if(!["admin", "root"].includes(currUser.username)) return res.status(403).json({ error: "Unauthorised user" });
 	try {
 		const user = await dbUtils.getItemById(User, userId);
     const clientData = user.toObject();
@@ -17,15 +19,17 @@ router.get("/:userId", async function (req, res) {
 		delete clientData.password;
 
 		res.json(clientData);
-	} catch (err) {
-		console.log(err);
-		res.json({ err });
+	} catch (error) {
+		console.log(error);
+		res.status(500).json({ error });
 	}
 });
 
 
-/* GET all categories. */
+/* GET all users. */
 router.get("/", async function (req, res) {
+  const currUser = res.locals.userData;
+  if(!["admin", "root"].includes(currUser.username)) return res.status(403).json({ error: "Unauthorised user" });
 	try {
 		const userItems = await dbUtils.getItemByField(User);
     const clientItems = userItems.map(item =>{
@@ -38,9 +42,9 @@ router.get("/", async function (req, res) {
     })
 
 		res.json(clientItems);
-	} catch (err) {
-		console.log(err);
-		res.json({ err });
+	} catch (error) {
+		console.log(error);
+		res.status(500).json({ error });
 	}
 });
 
@@ -50,7 +54,8 @@ router.get("/", async function (req, res) {
 router.put("/:userId", async function (req, res) {
 	const userId = req.params.userId;
 	const modifiedUser = req.body;
-
+  const currUser = res.locals.userData;
+  if(!["admin", "root"].includes(currUser.username)) return res.status(403).json({ error: "Unauthorised user" });
 	try {
 		const user = await dbUtils.updateItem(
 			User,
@@ -63,31 +68,34 @@ router.put("/:userId", async function (req, res) {
 		delete clientData.password;
 
 		res.json(clientData);
-	} catch (err) {
-		console.log(err);
-		res.json({ err });
+	} catch (error) {
+		console.log(error);
+		res.status(500).json({ error });
 	}
 });
 
 /* DELETE user by id. */
 router.delete("/:userId", async function (req, res) {
 	const userId = req.params.userId;
+  const currUser = res.locals.userData;
+  if(!["admin", "root"].includes(currUser.username)) return res.status(403).json({ error: "Unauthorised user" });
 	try {
 		const user = await dbUtils.deleteItemById(User, userId);
 		const clientData = user.toObject();
 		// delete the user's password
 		delete clientData.password;
 		res.json(clientData);
-	} catch (err) {
-		console.log(err);
-		res.json({ err });
+	} catch (error) {
+		console.log(error);
+		res.status(500).json({ error });
 	}
 });
 
 /* reset the user collection. */
 router.delete("/", async function (req, res) {
-	const newUser = require("../database/testData");
-  
+	const newUser = require("../database/defaultData");
+  const currUser = res.locals.userData;
+  if(!["admin", "root"].includes(currUser.username)) return res.status(403).json({ error: "Unauthorised user" });
 	try {
 		const userArr = await dbUtils.resetItems(User, newUser);
     const clientDataArr = userArr.map(user => {
@@ -98,9 +106,9 @@ router.delete("/", async function (req, res) {
     })
 
 		res.json(clientDataArr);
-	} catch (err) {
-		console.log(err);
-		res.json({ err });
+	} catch (error) {
+		console.log(error);
+		res.status(500).json({ error });
 	}
 });
